ppsi engine verbosity log optimized with new Kconfig PPSI_RUNTIME_VERBOSITY opt
This patch introduces a new Kconfig option, called PPSI_RUNTIME_VERBOSITY, which
disables runtime modification of ppsi verbosity, when it is not desired.
This brings a benefit in executable size, since all the
if (pp_diag_verbosity > ...) conditions can be evaluated at compile time (see
PP_PRINTF definition in ppsi).

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>