Fix error in flash-write
There was an error in flash-write.c, the buffer to write was not
being incremented correctly after a page was written, resulting
in the same 256 bytes (the ones at the start of the input pipe)
written to each page in the flash.

The error is now fixed and we're writing properly.

Signed-off-by: Theodor Stana <t.stana@cern.ch>