flash-write: massaging flash data buffer
There was somewhat of a bug in the spec_write_flash function of
flash-write. The memcpy function was not functioning 'as expected',
so I changed the fixed-size buffer for a pointer and malloc-ed it
to the size. This seems to fix the bug.

The issue about reading the flash status register has yet to be
fixed.

Signed-off-by: Theodor Stana <t.stana@cern.ch>